s3-scripts/requirements.txt
boto3==1.34.82
botocore==1.34.82
python-dateutil==2.9.0.post0
humanfriendly==10.0
